fix(coderd/provisionerdserver): address review on module file download fix
Add a debug-level log before the org-isolation rejection so operators have a
breadcrumb without leaking file existence to the calling provisioner. Move the
DownloadFile regression test into provisionerdserver_test.go and rename it to
TestDownloadFile to match the existing per-RPC test naming.
