+++ coderd/provisionerdserver/provisionerdserver.go
@@ -1596,6 +1596,10 @@ func (s *server) DownloadFile(request *proto.FileRequest, stream proto.DRPCProvi
 			return fail(xerrors.Errorf("authorize module file: %w", err))
 		}
 		if !ok {
+			s.Logger.Debug(ctx, "module file download rejected: file not referenced by any template version in daemon org",
+				slog.F("file_id", fid),
+				slog.F("organization_id", s.OrganizationID),
+			)
 			// Use the same error as the metadata check above so the handler
 			// does not confirm the existence of files in other organizations.
 			return fail(xerrors.Errorf("file %s is not a modules file", fid))
